backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.26.0
<<<<<<< Updated upstream
pyyaml>=6.0
=======

